Fixed a lot of errors

+++ Repositories/RoomRepository.py
@@ -13,37 +13,41 @@ class RoomRepository():
 
     def findUsersAtRoom(self, id):
 
-        return (RoomUser.select(RoomUser.user_id, User.username, Room.admin_id)
-            .join(User,attr='user')
-            .switch(RoomUser)
-            .join(Room,attr='room')
-            .where( (Room.id == id) ))
+        return (RoomUser.select(
+            RoomUser.user_id, User.username,
+            Room.admin_id).join(User, attr='user').switch(RoomUser).join(
+                Room,
+                attr='room').switch(RoomUser).where((Room.id == id)
+                                                    & (RoomUser.isInRoom == 1))
+                )
 
     def findPendingUsers(self, id):
 
-        return (RoomUser.select(RoomUser.user_id, User.username)
-                .join(User, attr="user")
-                .switch(RoomUser)
-                .where( (RoomUser.isInRoom == 0) & (RoomUser.room_id == id)))
+        return (RoomUser.select(RoomUser.user_id, User.username).join(
+            User,
+            attr="user").switch(RoomUser).where((RoomUser.isInRoom == 0)
+                                                & (RoomUser.room_id == id)))
 
     def isRoomAdmin(self, room_id, user_id):
 
-        return (Room.select(Room).where( (Room.id == room_id) & (Room.admin_id == user_id) ))
+        return (Room.select(Room).where((Room.id == room_id)
+                                        & (Room.admin_id == user_id)))
 
     def findRoomMessages(self, id):
-        return (Message.select(User.username, Message.content,
-                               Message.file_name, Message.file).join(
+        return (Message.select(User.username,
+                               Message.content).join(
                                    User,
                                    on=(Message.from_user_id == User.id),
                                    attr='user').where(Message.room_id == id))
 
     def rejectUser(self, user_id, room_id):
-        RoomUser.delete().where(
-            RoomUser.user_id == user_id & RoomUser.room_id == room_id).execute()
+        RoomUser.delete().where((RoomUser.user_id == user_id)
+                                & (RoomUser.room_id == room_id)).execute()
 
     def acceptUser(self, user_id, room_id):
-        RoomUser.update(isInRoom = True).where(
-            (RoomUser.user_id == user_id) & (RoomUser.room_id == room_id)).execute()
+        RoomUser.update(
+            isInRoom=True).where((RoomUser.user_id == user_id)
+                                 & (RoomUser.room_id == room_id)).execute()
 
     def saveRoom(self, room):
         return Room.create(**room)
